requirements.txt
# Core

langchain>=0.3.0
langchain-core>=1.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=1.2.0

# LLM Providers
langchain-groq>=0.2.0
groq>=0.11.0

# Vector Databases
langchain-chroma>=0.2.0
langchain-qdrant>=0.2.0
langchain-weaviate>=0.1.0
lancedb>=0.15.0

# Embeddings
sentence-transformers>=3.0.0

# Sparse Retrieval
rank-bm25>=0.2.2

# Evaluation Frameworks
ragas>=0.2.0
deepeval>=1.0.0

# Data
datasets>=3.0.0
pandas>=2.0.0
numpy>=1.24.0

# NLP Utilities
langdetect>=1.0.9

# Environment
python-dotenv>=1.0.0